requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-environ==0.9.0
django-redis==5.4.0

uvicorn==0.38.0


channels==4.1.0
channels-redis==4.2.0


whitenoise==6.8.2


mysqlclient==2.2.0

django-templated-mail==1.1.1

# JWT Authentication
djangorestframework-simplejwt==5.3.1


# CORS
django-cors-headers==4.3.1
